Move SecurityContextExtensions to api module
